requirements/app.txt
galaxy.plugin.api==0.30
requests==2.21.0
psutil==5.6.1
pyobjc-framework-Quartz==5.2; sys_platform == 'darwin'